Only sort PLIST files when there is no @command in-between
The previous idea of sticking unrecognized lines to the previous line
was too complicated and did not match existing practice (devel/blosc,
comms/xtel).

Now the $ is sorted like an ordinary character.

PLIST files that have an @command line somewhere in the middle are not
sorted because that exact position might be intentional.