Refactor autofix
The code for fixing the lines automatically was scattered around
several source files and not easily understandable. Now it lives in a
single file, in a single type even, and all the operations are done as
simple method calls on that type.

This refactoring prepares the --include command line option, which will
apply a filter to the diagnostics, allowing fine-grained control over
the exact diagnostics. This feature was not possible with the old code
since then every call would have been much more complicated, probably
adding another nesting level.