Autofix: enforce consistent argument order
This is necessary since the upcoming --only option will decide whether
to autofix something or not based on the diagnostic, and that is only
possible when the diagnostic is already known at that point. Alterna-
tively applying the actions could be delayed until Apply is really
called, but that would result in too much code.