Fix PLIST sorting and duplicate removal
A conditional entry is removed if there is an unconditional entry for
the same path.

An unconditional entry is only removed if another unconditional entry
for the same path appears earlier in the file.

PLIST sorting is first by path and then by the name of the conditional.
Before, it had been only by path, leaving the conditionals unsorted.