Remove Pkglint.CurrentDir
This variable had been confusing all the time, most importantly because
the "directory that is currently checked" does not correspond to
anything in real pkgsrc.

Those places that need a reference to that directory use the directory
of the current package now, which in most cases is the same as before.
Some of the variable checks don't make sense outside of a package, so
these have been disabled when checking individual files instead of a
whole package.

Some tests had to be rewritten to rely on the file system, and since
pkgsrc works heavily with files and their relationships, this is more
realistic than before.