Refactor loading data from files
Previously, it was hard to remember which of the loading functions did
exactly what. By unifying them into only two functions and passing the
variants as options, the code is more expressive now.

In some places, pkglint is now stricter than before. Especially when
loading files from the infrastructure, empty files are not valid
anymore. This only affects the tests, though.