Fix relative paths in diagnostics
Since a few days, pkglint had taken all paths relative to the pkgsrc
root. Due to this, files in the current directory were written as
../../category/current-package/filename.

To make this change work, the Package type had to get back its directory
field, from which all relative file paths are resolved. The PKGPATH
variable is now determined inside the NewPackage constructor function,
which means that all callers had to be adjusted. And there were a lot.

Many tests failed because of this change, and most of them have been
fixed with the minimal code change. To provide test coverage for really
relative paths (like running "pkglint ." or "pkglint category/pkgbase",
some of the tests are now run after changing to the temporary directory.