When tokenizing shell program, replace regular expressions with code
After the preparation of the last 15 or so commits, this final change
was finally possible. The change doesn't get any smaller than this, I'm
glad that all existing tests run.

The idea in the intqa package to have a direct unit test for each and
every type, method and function will prove useful in the future. Without
this test coverage, big changes like these are impossible to get right.
That's why several new tests have been added in the previous commits.