Implement line-buffered output
Since the recent rewrite of the SeparatorWriter, pkglint wrote the
diagnostics one byte at a time, which had a significant overhead because
of all the system calls.

This slowdown showed up prominently in the profiling data.

Using line-based buffering instead of block-based buffering is a
compromise between execution speed and convenience.