Avoid some memory allocations, just for fun
I first tried to replace the []string with [...]string, but that
resulted in exactly the same code. The explanation was generated each
time even though it would not be necessary in most cases.

The next approach was to put the allocation in an inner function. This
was more successful since that function is only called when the
explanation is actually needed. Because the function call is completely
inlined, this adds a bit to the code size but doesn't not incur any
run time penalty.