Rename Pkglint.CheckDirent to Pkglint.Check
Since this is the main entry point for the tests and for the each
command line argument, it makes sense to have a clear and simple name
for this method.

Before this renaming I had constantly been confused about how
CheckDirent and CheckFile work together. Now it slowly becomes clear.