Refactor Pkglint.checkDirent and checkReg
It had been confusing that checkReg actually checked whether the file
was a directory. Based on its name, this could never be the case.

Now the hierarchy among the methods is cleaned up:

Pkglint.Check
  Pkglint.checkDirent
    Pkglint.checkReg
    Pkglint.checkdirPackage
    CheckdirCategory
    CheckdirToplevel

During the execution of checkdirPackage, checkDirent is called for each
directory entry in the package, which makes sense based on the method
names.