Clean up code in VartypeCheck
This commit concludes the paper review of the whole pkglint code that
started at 2018-10-13. During this time, I have looked at every line of
the pkglint code and checked whether it is still current, follows a good
programming style and is generally understandable.

The review took quite long, but looking back, it was very efficient.
During the review I found many bugs for edge cases. Some of them I fixed
right away, and some I left for later by just marking them in the code
with the usual TODO and FIXME comments. There are over 250 of them.

Fixing them all will take more time but now these bugs are at least
known and documented.

During the review I also added lots of documentation to the code, which
should make it easier for other developers to contribute to pkglint.