Add tests for 100% branch coverage in options.go
The handled != nil condition was completely redundant, as well as the
declared != nil in the paragraph above it. The branch coverage could
only detect the lower instance though because of the order of the
conditions.

To point out the symmetry between these cases (and of course for
performance), both cases have been aligned to each other.