Make explanation for variable permissions more human-readable
Saying "any file" instead of a simple "*" is clearer to human readers
but makes it more difficult for computers to parse. This is no deal
since the explanations are solely targeted at humans.

Using the Oxford comma for the possibly long lists proves useful for
list items consisting of multiple words (like "set a default value") but
doesn't feel right for single-word list items like "set, or used".
Most probably I should look up the exact rules for the Oxford and the
Cambridge comma again.