License is not a list variable
It is allowed (though often questionable) to append to a license
variable. It such a case, the initial value of LICENSE is typically
a simple license name. To that, it is possible to append and "AND b"
or "OR b" without introducing syntax errors.

It is not possible to extend an already complex license condition in the
same way. For example the expression "a AND b" can only be extended with
another "AND c", but not with an "OR c" since the operators on each
level must be the same, i.e. either all AND or all OR.

The LICENSE variable is therefore not a list of things.