In qa-pkglint, use *.ref as reference file instead of reference.out
This way, the basename of the reference file still refers to the Git
commit and to the other related files. This makes it easier to see how
to regenerate that file.

The implicit assumption is that there is only ever one *.ref file. This
is guaranteed since I'm the only one using this program.