In the check for distfile hashes, store only the location
Before this commit, the final heap dump from running pkglint on both
pkgsrc and pkgsrc-wip was 117 MB, after this commit it is 101 MB.

For reporting mismatches between distfiles, the text of the distinfo
line is not needed. The only information necessary is the location of
the line and the previous hash. The location is now stored separately
from the text of the line, which already results in 16 MB saving.

The full text of the distinfo files is still contained in the heap dump
since the hash values are substrings of the whole file content.

The number of garbage collections increased from 414 to 446 because the
overall size of the heap is now less than before. The total amount of
memory freed by the GC stays around 18.6 GB.