Intern the license names
Without this simple change, the whole text of the file containing the
license name is kept in memory. An example for this waste of memory is
mail/sendmail/Makefile.common. That whole file is contained in the final
heap dump just because the word "sendmail-license" is a key in the map
Pkgsrc.UsedLicenses.