Disable check for expanded CVS Ids in pkgsrc-wip
It's only a note, not a warning, but it appears more than 10000 times.
As soon as a package is re-imported into main pkgsrc, these CVS Ids will
be overwritten anyway, so it's not a big deal. Sure, they are not in a
canonical form, but essentially nobody relies on them anyway.