Remove dead condition in parsing continuation lines
At this point there cannot be a completely empty textnl. In most cases
it will be some string terminated by "\n". And even if that's not the
case, the line cannot be completely empty since then it would not have
been parsed as a line but merged into the previous line.