Remove redundant condition from MkLine.VariableNeedsQuoting
When a variable appears as a whole word, that word cannot appear in
quotes at all. Otherwise the quotes would delimit the word, and whatever
is inside the quotes is considered part of that word already.

The current code allows for edge cases like "before ${VAR} after", which
on the first sight looks as if the variable would be its own word, but
that code is wrong.