Reverse the two lines in the note for redundant variable definition
In the usual case, the first line comes from the included file while the
second line comes from the including file. The including file is more
specialized, therefore it is easier to remove the redundant definition
there.

This doesn't work when a variable is first defined in a Makefile and
later defined to the same value in an included file. This needs to be
fixed.