Split the distinfo check into parsing and the actual checks
This changes the order of the diagnostics so that they don't appear in
the natural order (from top to bottom) anymore. It's not perfect but
acceptable since distinfo files are only rarely written manually.

The diagnostics for mismatched hash algorithms are now reported at the
first line of the group instead of one line after the last line since
that makes more sense. When an editor jumps to the given line number,
the following lines are usually visible. This also removes the number of
EOF lines in the diagnostics, which most editors also don't understand.