Restructure and document the distinfo checker
This is in preparation for adding the missing hashes for distfiles.
Currently there are 248 distfiles in pkgsrc that are still missing the
SHA512 hash. Several of these distinfo files combine the distfiles for
more than one platform. Running "bmake distinfo" on these packages
discards all hashes except for the currently selected platform.

To support these packages, pkglint will detect whether the distfiles
have been downloaded and autofix the missing hashes, provided that the
current hashes match the downloaded distfile.