Make redundancy check more precise
When a default assignment has no effect, it is clearly marked as such.
Before, it was marked as "redundant", which was ambiguous.

The redundancy check is more precise because it only reports cases where
the variable value is guaranteed to be constant. Before, there might
have been cases where appending to a variable would lead to wrong
warnings.