Fix relpath
I think I got a much better idea now about the purpose of the relpath
function. This had been unclear for a long time.

Relative paths in pkgsrc should always be specified by going to the
pkgsrc root first, and from there to the destination. This keeps the
relative paths constant if a package should ever be moved into a
different category (which has not happened in the last 20 years).