Fix check for GNU_CONFIGURE and USE_LANGUAGES
Using just the first (or last) definition of a variable is not correct.
There are too many cases that are more complicated in practice. The
recently added Var type provides a ground allowing checks that are more
reliable than before, by taking into account that a variable value may
be assembled in multiple places to form a final value.