Prepare for refactoring MatchVarassign
It is too complicated to analyze the failing high-level tests to see the
actual differences at the character level. Therefore run the old and new
implementations in parallel, immediately marking all differences.

For now, the old and new implementations are exactly the same. As said
above, it's only the preparation.