Split parsing of Makefile lines into more phases
In the first phase, the line is split into a main text and the comment.
This is where unescaping of the # takes place, uniformly for all lines
except those starting with a tab.

After that, the # does not have a special meaning anymore, which makes
it an ordinary character.

The bmake parser works the same.