Add systematic tests for the redundancy check
That part of pkglint is so new and so complex that it needs thorough
testing. In the past there have been really many little quirks and big
obvious mistakes.

Combining all these tests sounds like a lot of work. Investing this time
is probably worth every minute because of the various interactions and
edge cases.