Move warnings for redundancy check into RedundantCheck
The loose coupling between the redundancy check and the effect of
generating warnings had been one abstraction level too much. Since the
pkglint code is only used by a single program right now it can be as
concrete as possible. If anyone should ever want to use pkglint as a
library to implement further checks, the API can still be changed.