Flatten RedundantScope.handleVarassign
In addition, the redundant check for len(writeLocations) has been
removed.

There is an early return for the very first variable access. There is
also an early return for a variable that has only ever been read. This
combination assumes that a variable can only be read and written. Since
this was non-obvious, the code has been rewritten to test the write
locations directly.