Adjust variable permissions to real-world usage
Since the redundancy check is now in action, the variable permissions
don't need to be that strict anymore. They don't need to distinguish
between assigning and appending and what is allowed in which files since
that job is done by the redundancy check.

This commit already covers the majority of the 10000 lines patch for the
whole pkgsrc tree. The remaining parts will not contain that many
repetitions and be handled in the next commit.

The pkglint unit tests are completely unaffected by this change. This
means these warnings have to be tested in the real pkgsrc tree, which is
efficient and reaches the goal of documenting common practice.