Allow pkglist variables in Makefile.* in addition to Makefile.common
Although these files have grown historically and are mostly
undocumented, there are still many of them in pkgsrc. To help package
authors migrate to documented *.mk files, pkglint needs to issue more
specific warnings and provide assistance in migrating. The permission
warnings are not the appropriate tool for achieving this.