Splitting Makefile lines on whitespace is no longer necessary
Since 2015 the .for loop splits in exactly the same way as the :M
variable modifier. Therefore it only makes sense for pkglint to split
variable values in the same way.

Right now pkglint differs in the way it treats backticks. It considers
`pkg-config arg1 arg2` a single word, while bmake splits it into three
separate words. For analyzing CFLAGS this is a beneficial side effect,
in other cases it may prove wrong in the future to implement field
splitting _almost_ in the same way as bmake.