Improve check for compiler and linker flags
The previous version of this check issued wrong warnings for several
flags pkglint didn't know. This gave a bad impression, but nobody has
complained so far. The warnings were not useful at all.

The new version focuses on real problems and education instead. It
distinguishes compiler and linker flags and warn if the flags are
assigned to the wrong variables.