Only load regular files when scanning mk/ for untyped variables
This fixes Test_Pkgsrc_loadUntypedVars__badly_named_directory on NetBSD,
which allows directories to be opened as if they were regular files,
resulting in binary data.

It would be more thorough to check for regular files in Load, but that
would cost another system call per loaded file, which would be too slow.