Completely rework the variable alignment check
Doing the check on the parsed and joined lines was wrong from the
beginning and made the code more complicated than necessary. It was also
too difficult to fit all the realignment methods into a single method.
Splitting them up into the five different cases makes each of them
manageable. This in turn is important since there are still several
FIXME markers in the tests for this check.

A few of these markers have already been fixed, but the majority of them
is still waiting for the next commits.