Remove redundant indentation check for top-level Makefile
The normal check for aligned variables is good enough. That check is
getting more stable and predictable now, and since it can be autofixed,
it will become a standard check soon (no need for -Wspace).

The goal behind removing this check is to eliminate MkLine.ValueAlign,
since that method has been made redundant by VaralignBlock.