Add checker for _VARGROUPS
Several files in the pkgsrc infrastructure are quite big and complicated
to understand, such as mk/compiler/gcc.mk with its more than 100
variables.

The _VARGROUPS section tries to at least organize this complexity a bit,
to make it more accessible. It comes with a burden though, in that it
adds a lot of redundancy. All these redundant definitions need to be
crosschecked, to ensure they are free from simple mistakes.