Make conditions in Autofix logging simpler
The first two are obvious to simplify.

The condition around the Separate call was redundant because the
SeparatorWriter takes care internally to only add a separating empty
line if there hasn't been an empty line before. If an explanation had
been written before, there is exactly such an empty line, therefore the
call to Separate didn't have any effect.