Fix Tester.SetupPackage
The lines of the package Makefile are filled up so that the custom
lines always start at 20. Before, they were filled up with "# filler"
comments, which meant that the custom lines were in the same paragraph
as the .include line. This in turn prevented the check for variable
value alignment and confused the rationale detection.