Add Path type to avoid using plain strings for this purpose
In pkglint, there are quite different kinds of paths in use:

Paths using the separator character from the operating system are only
used when parsing the command line. These are translated to slash-
separated paths as fast as possible.

Paths relative to the current working directory are used in the
diagnostics.

Paths relative to the pkgsrc root are used in the PKGPATH variable,
like graphics/ImageMagick.

Paths relative to the pkgsrc package directory are used to refer to
other packages, like ../../category/package.

All these kinds of paths cannot be combined arbitrarily. As a first
step, pkglint now distinguishes between strings and slash-separated
paths. This is similarly useful as the decision to use backticks only
for regular expressions, and double quotes otherwise. And of course
the decision to represent regular expression strings as regex.Pattern.