Use distinct Path type for slash-separated paths
In this commit, the calls to hasPrefix and hasSuffix were translated
into the corresponding HasPrefixText and HasSuffixText, even if it were
more sensible to use HasPrefixPath and HasSuffixPath. This is to
preserve the current behavior. All these places need to be inspected and
converted manually, along with writing the necessary tests to highlight
the differences.

In the first try of this commit, there was a single package whose
diagnostics changed: wip/clfswm. Since I was unsure which of the inexact
HasPrefixPath/HasSuffixPath changes caused this change, this commit
preserves the previous behavior.