Remove unnecessary code in Package.load for skipping FILESDIR
The intention of that line was to skip files from FILESDIR, since any
Makefile that is placed there does not have to be a BSD-style Makefile.

The -Cextra command line option had been removed several weeks ago,
therefore this check became unnecessary and should have been caught by
the branch coverage analysis.

That plan didn't work though because the line was sloppily worded.
Instead of testing whether the file _is in_ the FILESDIR, the test was
whether the patch contains "files/" somewhere, and for sure, the path
"x11/xorg-cf-files/Makefile" contains that string, which means that
conditionally included files were not checked properly for that package.