Remove Autofix.Anyway
It was not needed at all. All the existing uses of the Autofix feature
already guarded themselves properly to only warn if something was
actually wrong. Therefore, each diagnostic that is given to an Autofix
is actually logged.

The effect of removing Autofix.Anyway (and enabling it by default) is
that some more diagnostics are logged. Previously, diagnostics that
applied more than once to a single line were suppressed because pkglint
was unsure which of them shouldd be autofixed. Now these are logged as
well, even if they are not autofixed.