Move relpath to pkgsrc.go since it is tightly coupled with Pkgsrc
The relpath function computes a relative path by first going to the top
directory of pkgsrc and from there to its destination path. Therefore it
is appropriate to express this dependency when calling the function,
instead of relying on a global variable.

In the next commit, the function will be converted into a method of the
Pkgsrc type.