Fix Pkgsrc.Relpath to return better relative path
It's funny that replacing the questionable code with the correct code
immediately fixes the problem, even if I didn't see it before.

A run over the complete pkgsrc directory confirms that the change is
good. Several warnings that had longer paths before are now shorter and
more to the point.