In Tester.SetUpHierarchy, use an explicit base directory
During a refactoring that added more specialized path types, I noticed
a type conflict. The expression arg.lines.Filename is already relative
to the current directory and therefore does not need to be passed
through Pkgsrc.File.

To better understand the situation, the code dealing with the relative
paths is now in a separate function.