Remove command line option -Wspace
The check for alignment of variable assignments is good enough that it
can be enabled by default. Most of the alignment issues have been fixed
by a few big commits all over pkgsrc.

This change of the default behavior produces around 3000 new notes. Most
of them are about patches that are missing documentation anyway. As
these can be automatically fixed by pkglint (including the necessary
changes to the distinfo file), it shouldn't be too discouraging.