Demonstrate that loop variables can end up in global variables
This produced a hard to understand warning when the variable f was used
in a .for loop. "To use f at load time" simply does not apply here since
f is a load-time-only variable.

~~~shell
pkglint -d git/pkgsrc/ 2>&1 \
| grep Untyped \
| grep -v 'Untyped variable "[A-Z_][A-Z0-9_]*"' \
| sort \
| less
~~~

The resulting list contains many variable names that are clearly
invalid. Among them are the empty name, a slash, names starting with a
+, names ending with a comma, and so on.

This part of pkglint needs a big cleanup.