Make the QA test independent from check.v1
Just because it works. And to demonstrate that the QA check really does
not need anything from the Suite type, which has grown quite large.

A side effect of this change is that the QA test is always performed,
even when the check.v1 tests are filtered. This was unexpected at first,
but is useful.