Remove simulateSubstLines for testing
When this function was added some years ago, MkLines.ForEach didn't
exist. Back then it was a good choice to use an abstraction layer to
avoid having the same test fixture in several tests.

Now, with MkLines.ForEach, the test fixture is easy to set up, therefore
there is no need anymore for the additional abstraction layer.

The simulateSubstLines function also made it slower to step through the
code since it added an additional 9 empty lines at the beginning of the
file. This was just done to have the same number of digits in all the
line numbers, for aesthetic reasons.