Fix handling of conditional assignments in SubstContext
The previous logic of detecting missing variable assignments in
branches of conditions was completely broken and didn't even work in
simple cases like having no variables set in a single branch.

It has been rewritten completely, based on an appropriate definition of
a conditional branch and its components. The new data structures will
allow to pinpoint the exact branch of the missing lines.