Start prototyping the variable assignment alignment
I am unhappy with the current code because it has too many special cases
and gets some corner cases wrong. It is hard to understand because the
details are not nicely abstracted away.

The aim of the next implementation is to model the indentation based on
the actually observed rules. When I wrote the specification in November
2019, the wording made sense, and it felt like a vast improvement over
the current code, which was built without a detailed specification.