Add test for different right margins in same paragraph
Before commit bf0a803e7e45 on 10.12.2019 at 21:57 +0100, this test
failed since up to then, the right margins of all MkLines in a paragraph
were aligned to the same column. There wasn't a single unit test for
this case though.

In order to better understand the small effects of fixing the layout of
these variable assignments, it's ok to have the current smaller scope
for the right margin.