Make the API for varalignMkLine.rightMargin simpler
Previously, the main code called rightMargin with a different argument
than the test code. This was unnecessarily complicated and made the
tests misleading.

The tests have been adjusted and a few new tests have been added to show
the differences between MkLines with and without a value in their first
line.