Clean up tests for subst
The filename is only used for the diagnostics, therefore it is better to
use the default filename, as documented in Tester.NewMkLines.

The MkCvsID lines are not necessary for the subst tests since that check
is specific to MkLines.Check. It was necessary when these tests called
MkLines.Check, not long ago.