Ensure literal identifiers for SUBST_CLASSES and other variables
There are several places in pkgsrc where it only makes sense to use
direct identifiers that are not based on other variables. And there are
also places where referring to other variables makes sense.

Therefore, the existing type Identifier has been split into
IdentifierDirect and IdentifierIndirect, just like in YesNoIndirect.

The decision whether to allow indirect identifiers for a specific
variable has been made based on the current practice in pkgsrc. It may
prove to be wrong in a few edge cases. For user-settable and
system-provided types, it doesn't matter anyway since these variables
are not supposed to be defined at all by pkgsrc packages.