Fix comment on SubstContext.isConditional
Starting the scope on the first dependent line has the unwanted side
effect that the scope would end when the conditional is closed, even
though it is legitimate to continue the remaining variables at the same
level as the SUBST_CLASSES addition.

This is the same as the lexical scoping rule used by many programming
languages.