Rewrite SUBST handling to properly check lexical scoping
Before, pkglint had been confused when a SUBST_CLASSES block was split
over several paragraphs. To fix this, pkglint now uses the typical
lexical scoping rule that is familiar from many other programming
languages.

Adding an identifier to SUBST_CLASSES declares that identifier in the
current scope. Scopes are nested using the .if and .endif directives.
The .for and .endfor directives are completely ignored since they don't
make sense in the context of SUBST classes, as the SUBST IDs have to
be literal identifiers.