Fix Test_MkLineChecker_checkShellCommand__indentation
This test called MkLines.Check twice on the same object, which is not
realistic.

The next commit will change Autofix to always modify the fixed lines in
memory, even if neither of the --show-autofix nor --autofix options is
given.

The first call to MkLines.Check (without --autofix) would therefore
already fix the internal text. The second call (with --autofix) would
then see the fixed text and would not generate the AUTOFIX diagnostics.