Add more test cases for varalignLine.alignValueMultiFollow, part 2
I was still trying to find an example where tabWidth was wrong and would
lead to different results than tabWidthAppend, until I noticed that
this is a follow-up line, which means that there cannot be a varOp part
in that line, which makes spaceBeforeValue equivalent to the whole
beginning of the line.

There might be a difference though if the line has no value but only a
continuation backslash.