Remove unused Autofix.ReplaceRegex
It was a nice idea, but in the end the code for managing the repetitions
was too complicated to be maintained any longer. Those places where this
method has actually been used didn't make use of all the available power
that came by using regular expressions in this part.

Especially when modifying the text automatically, it is better to be as
specific as possible and not use any patterns or placeholders.