Ensure that MkLines.checkAll is only called once
This prevents long debugging sessions because the first call modifies
the mklines data and the second call assumes a fresh state. This happens
especially when using Tester.ExpectDiagnosticsAutofix.

As a byproduct, this assertion made it necessary to fix another bug that
had gone unnoticed and affected packages that use PKGDIR to refer to
another package's directory.