Fix resolution of variable value in package
Before, pkglint was really careless about the variables scope of the
package. It had just added whatever variable definitions it saw in any
Makefile fragment that was loaded while the package was checked.

Now only those files that are included by the package Makefile can
contribute to the variable scope of the package. This makes much more
sense.