Add --network option to check for reachable HOMEPAGE
This will be used to fix the wrong SourceForge URLs that had been
generated by the previous pkglint commits. The plan is to also toggle
the homepage URLs between http and https, whichever is reachable.

This commit discards the "strict pkglint no-network-access policy" and
weakens it to "only access the network on explicit demand", which is a
reasonable compromise between useful features, quick run time and
resource use within the expected limits.