Add HomepageChecker.isReachable
This will be used for deciding whether a homepage can be migrated from
http to https. That migration only makes sense if the https page is
indeed reachable. For some URLs this test is not necessary since they
are well-known to support https, such as GitHub.

The code is similar to checkReachable but still much simpler, therefore
it doesn't make sense to merge these two functions.