Accept "unknown network errors" for Go 1.12
The possibility to Unwrap errors was added in Go 1.13. Earlier versions
would have to use a type switch with all known error types to unwrap the
root cause. Implementing would be too much unnecessary work for pkglint
since it only needs to run on Go 1.13.

This should fix the build problems on Linux with Go 1.12. It took quite
a while to implement such a simple "is this URL reachable" check using
Go because of the undocumented API additions and the overly convenient
ListenAndServe method that makes race conditions really simple. Maybe I
should have looked at the httptest package, which seems designed for
exactly my use case of having a dummy HTTP server in a unit test.