Add test case for version parsing
In a version number where the "nb" is directly adjacent to a letter, it
is not obvious how the version is parsed. It is parsed as expected.

Just good that none of the special-value words like pre or alpha ends
with an "n" or "b". But wait, "beta" starts with a "b" that might be
interpreted as the end of "nb". That will be the next test case.