Make Path.IsAbs faster
In a pkglint run that took 673 seconds, 37 seconds were used by the
method Path.IsAbs. Most of that time was spent in strings.Replace, in
order to replace forward slashes with backslashes. This conversion was
only intended for Windows, and even there it is not necessary since in
Windows paths, both slashes are treated as equivalent.