Only list packages in category SUBDIR
It does not make sense to list non-package subdirectories since they are
ignored anyway.

Noticed by looking at cross/Makefile, which has a comment that nios2
contains only helper files. This comment is unnecessary, but was
required previously, just to appease pkglint.

This aligns the category check with the top-level check, which since
about forever had the requirement that only subdirectories containing
a Makefile are listed there.