Next try to optimize Scope.DefineAll
The previous try failed since the argument to Scope.DefineAll was not a
pointer but a copy of the other scope.  Building a cache on the copied
object obviously didn't have a big effect.  The code ran faster by a few
percent.  Not sure whether it is a measuring error.