Check for contradictory conditions in separate lines
This requirement made it necessary to completely rewrite the previous
check.  The check cannot be anchored at an "And" node since that would
skip conditions that consist of a single term only.

Instead, all conditions that are guaranteed to be satisfied in the
current line are collected.  This also includes the conditions from
previous lines in the same file.  Conditions in double negation are
ignored for now since these don't occur in practice.  Conditions in
parentheses or in complex expressions are handled correctly though.

The only facts that are collected are pattern matches with the :M
modifier.  The :N modifier is used less often and is therefore ignored,
for now.

Equality comparisons are ignored for now.  They could be added easily
though, as well as checks that span multiple variables.  For example,
OPSYS and MACHINE_PLATFORM are so closely related that it makes sense to
cross-check them.