Check that CFLAGS and similar variables are only appended
Using the assignment operator '=' can remove the default value of the
variable.  In contrast to the variables in mk/defaults/mk.conf, there
are a few variables that are already defined in sys.mk and thus loaded
at the very beginning.  Assigning them a value would overwrite the
default value even before bsd.prefs.mk is included.