Fix tests for distinfo hashes
In the previous commit, I hurried too much. Now look carefully again to
keep the purpose and intention of each test up-to-date.

The test some_algorithms_in_wrong_order was an exact duplicate of the
test directly above it, and it was an edge case anyway. When there are
only two real distfile hash algorithms, chances are low enough that any
of them are missing.