Extend explanation for documenting patches
About half of all pkglint error messages are about undocumented
patches. When using mkpatches to generate the patch files, these files
may be regarded as auto-generated (which is true) and thus not intended
to be edited manually (which is false). Try to work around this
misinterpretation by providing more context to where the documentation
should go in the patch file. The previous version of only having a CVS
Id line as context was too little to be easily recognizable.