Rename dependency pattern to package pattern
A package pattern matches a package independent of its location within
pkgsrc. It is used in conflict declarations as well as in
pkg-vulnerabilities.

A dependency pattern with path is used in dependency declarations. Since
the former "dependency pattern" has been renamed now, let the former
"dependency pattern with path" take its place.