Use specialized data types for one-time actions
This removes even the slightest chance of accidentally suppressing a
diagnostic just in case two diagnostics have the same 64-bit hash code.
It also makes debugging easier.

The Logger type is kept alive throughout the whole pkglint run, which
may span the complete pkgsrc tree. Since the OnceStringSlices type does
not keep references to external strings, it cannot accidentally keep the
whole content of some file in memory just because some substring of it
is stored in the map of seen strings.

Storing the actual strings instead of a 64-bit hash increases the
memory usage. Since the long-lived duplicate avoidance works on the
diagnostics that are actually output, the additional memory usage is
limited by that size, which for a complete pkgsrc scan amounts to less
than 20 MB.